# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: BGE Héroes de la Patria API
  description: |
    API REST del Sistema de Gestión Educativa BGE.

    ## Autenticación
    La mayoría de endpoints requieren autenticación JWT. Incluir el token en el header:
    ```
    Authorization: Bearer <token>
    ```

    ## Rate Limiting
    - API General: 100 requests/minuto
    - Autenticación: 5 intentos/5 minutos
    - AI Endpoints: 50 requests/hora

    ## Códigos de Error
    - 400: Bad Request - Datos inválidos
    - 401: Unauthorized - Token faltante o inválido
    - 403: Forbidden - Sin permisos
    - 404: Not Found - Recurso no encontrado
    - 429: Too Many Requests - Rate limit excedido
    - 500: Internal Server Error

  version: 4.0.0
  contact:
    name: Soporte BGE
    email: soporte@bge-heroes.edu.mx
  license:
    name: Proprietary
    url: https://bge-heroes.edu.mx/terms
servers:
  - url: https://bge-heroes.edu.mx/api
    description: Producción
  - url: http://localhost:3000/api
    description: Desarrollo
...
paths:
...
  /tournaments:
    get:
      tags: [Tournaments]
      summary: Listar torneos
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [upcoming, active, completed]
        - name: type
          in: query
          schema:
            type: string
            enum: [quiz, challenge, hackathon]
      responses:
        '200':
          description: Lista de torneos
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Tournament'
  /tournaments/{id}:
    get:
      tags: [Tournaments]
      summary: Detalle de torneo
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Información del torneo
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Tournament'
  /tournaments/{id}/register:
    post:
      tags: [Tournaments]
      summary: Registrarse en torneo
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Registro exitoso
        '400':
          description: Torneo lleno o ya inscrito
        '402':
          description: IACoins insuficientes para cuota
  /tournaments/{id}/leaderboard:
    get:
      tags: [Tournaments]
      summary: Clasificación del torneo
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Tabla de posiciones
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        rank:
                          type: integer
                        participant:
                          type: object
                        wins:
                          type: integer
                        score:
                          type: number

  # ========================================
  # MARKETPLACE
  # ========================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Tournament:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        tournamentType:
          type: string
        format:
          type: string
        entryFeeCoins:
          type: integer
        prizePoolCoins:
          type: integer
        maxParticipants:
          type: integer
        currentParticipants:
          type: integer
        status:
          type: string
        startsAt:
          type: string
          format: date-time
...